Rename folder for clarity
This follows a brief discussion in Slack a while ago where it
was agreed to group the minutes by competition cycle since the
Core Teams will change between cycles.

Since these minutes relate to the SR2019 season and the folder
contains minutes from both 2018CE and 2019CE the previous name
was confusing by all measures. The new name avoids that.